refactor: inline constants.ts
